requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
jinja2==3.1.4
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.7.0
